+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+## [0.3.0] - 2026-01-02
+
 ### Added
 
 - `--jobs` / `-j` option to limit the number of threads used for parallel processing.